refactor: share tour layout
